requirements.txt
python-dotenv
pandas 
numpy
pymongo
certifi
pymongo[srv]
scikit-learn
mlflow
pyaml
fastapi
uvicorn
python-multipart
jinja2

# -e .